+++ Assignment1/src/main/java/com/example/assignment1/service/PhoneNoService.java
@@ -1,5 +1,6 @@
 package com.example.assignment1.service;
 
+import com.example.assignment1.dto.PhoneDto;
 import com.example.assignment1.entity.PhoneNo;
 import com.example.assignment1.repository.PhoneNoRepository;
 import org.springframework.stereotype.Service;
@@ -7,32 +8,53 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 @Service
 public class PhoneNoService {
+    String message;
    private final PhoneNoRepository phoneNoRepository;
-
-    public PhoneNoService(PhoneNoRepository phoneNoRepository) {
+   public PhoneNoService(PhoneNoRepository phoneNoRepository) {
         this.phoneNoRepository = phoneNoRepository;
     }
 
-    public String savePhoData(PhoneNo phoneNo){
-        phoneNoRepository.save(phoneNo);
-        return "Data Successfully Save !!!!!!!!!";
+    // Method for saving data into repository or database;
+    public String savePhoData(PhoneDto phoneDto){
+       if(phoneDto !=null) {
+           PhoneNo phoneData = new PhoneNo();
+           phoneData.setPhoneNo(phoneDto.getPhoneNo());
+           phoneData.setZipCode(phoneDto.getZipCode());
+           phoneNoRepository.save((phoneData));
+           message = "Data Successfully Save !!!!!!!!!";
+       }
+       else
+       {
+           message="Please Enter Proper Data....Try Again!!!";
+       }
+       return message;
     }
-
+   // Method For Fetching data from repository
     public List<PhoneNo> getPhoData() {
         return phoneNoRepository.findAll();
     }
-
-    public String updatePhoData(PhoneNo phoneNo, int pId) {
-        PhoneNo obj =phoneNoRepository.findById(pId).get();
-        obj.setPhoneNo(phoneNo.getPhoneNo());
-        obj.setZipCode(phoneNo.getZipCode());
-        phoneNoRepository.save(obj);
-        return "Data Successfully updated !!!!!!!";
+    // Method for updating existing data
+    public String updatePhoData(PhoneDto phoneDto, int pId) {
+       message=null;
+        if(phoneNoRepository.existsById(pId)){
+            PhoneNo obj =phoneNoRepository.findById(pId).get();
+            obj.setPhoneNo(phoneDto.getPhoneNo());
+            obj.setZipCode(phoneDto.getZipCode());
+            phoneNoRepository.save(obj);
+            message="Data Successfully updated !!!!!!!";
+        }
+        else
+            message="You have entered wrong Id ?? Try Again !!!!!";
+        return message;
     }
-
+   // Method For Deleting the existing data
     public String deletePhoData(int pId) {
-        phoneNoRepository.deleteById(pId);
-        return "Data Successfully Deleted With Id"+pId;
-
+        message = null;
+        if (phoneNoRepository.existsById(pId)) {
+            phoneNoRepository.deleteById(pId);
+            message = "Data Successfully Deleted With Id" + pId;
+        } else
+            message = "you have entered wrong details ?? ....!!!!!!";
+        return message;
     }
 }
